# Use constants for time durations (2x faster, 3.5x less memory)