A small optimization. ;)

Benchmark:
```ruby
require 'benchmark/ips'
require 'benchmark/memory'
require 'date'
require 'active_support/core_ext/numeric'
require 'active_support/core_ext/integer'

require_relative 'optimized'
require_relative 'original'

from = Time.now
to = Time.now + 1.year + 2.months + 3.weeks + 4.days + 5.hours + 6.minutes + 7.seconds

[:ips, :memory].each do |type|
  Benchmark.public_send(type) do |x|
    x.report('optimized') { DOTIW::TimeHashOptimized.new(nil, from, to) }
    x.report('master') { DOTIW::TimeHash.new(nil, from, to) }

    x.compare!
  end
end
```

Results:
```ruby
IPS:
Warming up --------------------------------------
              master     4.040k i/100ms
           optimized     8.728k i/100ms
Calculating -------------------------------------
              master     41.795k (± 5.3%) i/s -    210.080k in   5.040932s
           optimized     90.519k (± 5.2%) i/s -    453.856k in   5.027942s

Comparison:
           optimized:    90518.6 i/s
              master:    41794.8 i/s - 2.17x  (± 0.00) slower

Memory:
Calculating -------------------------------------
              master     6.072k memsize (     0.000  retained)
                        91.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
           optimized     1.704k memsize (     0.000  retained)
                        35.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)

Comparison:
           optimized:       1704 allocated
              master:       6072 allocated - 3.56x more

```