@dblock 2020-11-01
NIce. Want to add a line to https://github.com/radar/distance_of_time_in_words/blob/master/CHANGELOG.md and check-in the benchmark code into benchmakrs/some-name-that-makes-sensee.rb?