@krzysiek1507 2020-11-02
Changelog entry and benchmark added!

I wonder if should we add `benchmark-ips` and `benchmarks-memory` to dev dependencies. It be used very rarely but CI and users would have to install it.