@dblock 2020-11-02
No strong opinions of whether to include those gems, those running benchmarks can add them.