@dblock 2021-03-18
This was released in 5.3.0.